Refactor unit tests and add TestUtils utility methods
