+++ src/Postmark.Tests/ClientSuppressionTests.cs
@@ -17,7 +17,7 @@ namespace Postmark.Tests
         protected override void Setup()
         {
             _adminClient = new PostmarkAdminClient(WRITE_ACCOUNT_TOKEN, BASE_URL);
-            _server = MakeSynchronous(() => _adminClient.CreateServerAsync($"integration-test-suppressions-{Guid.NewGuid()}"));
+            _server = TestUtils.MakeSynchronous(() => _adminClient.CreateServerAsync($"integration-test-suppressions-{Guid.NewGuid()}"));
             _client = new PostmarkClient(_server.ApiTokens.First(), BASE_URL);
         }
 
@@ -79,7 +79,7 @@ namespace Postmark.Tests
             Assert.True(suppressionResult.Suppressions.All(k => k.Status == PostmarkSuppressionRequestStatus.Suppressed));
 
             // Suppressions are being processed asynchronously so we must give it some time to process those requests
-            var suppressionListing = await PollUntil(() => _client.ListSuppressions(new PostmarkSuppressionQuery()),
+            var suppressionListing = await TestUtils.PollUntil(() => _client.ListSuppressions(new PostmarkSuppressionQuery()),
                 k => k.Suppressions.Count() == 5);
 
             Assert.Equal(5, suppressionListing.Suppressions.Count());
@@ -105,7 +105,7 @@ namespace Postmark.Tests
 
             // Suppressions are being processed asynchronously so we must give it some time to process those requests
 
-            var suppressionListing = await PollUntil(() => _client.ListSuppressions(query), k => k.Suppressions.Count() == 1);
+            var suppressionListing = await TestUtils.PollUntil(() => _client.ListSuppressions(query), k => k.Suppressions.Count() == 1);
 
             Assert.Single(suppressionListing.Suppressions);
 
@@ -116,33 +116,12 @@ namespace Postmark.Tests
             Assert.Equal("ManualSuppression", actualSuppression.SuppressionReason);
         }
 
-        private async Task<T> PollUntil<T>(Func<Task<T>> pollingTaskFunc, Func<T, bool> isComplete, int retriesLeft = 5, int delayInMs = 1000)
-        {
-            var result = await pollingTaskFunc();
-
-            if (isComplete(result) || retriesLeft == 0)
-            {
-                return result;
-            }
-
-            await Task.Delay(delayInMs).ConfigureAwait(false);
-
-            return await PollUntil(pollingTaskFunc, isComplete, --retriesLeft);
-        }
-
         private Task Cleanup()
         {
             return Task.Run(async () =>
             {
                 await _adminClient.DeleteServerAsync(_server.ID);
             });
-        }
-
-        private T MakeSynchronous<T>(Func<Task<T>> t)
-        {
-            var f = Task.Run(async () => await t().ConfigureAwait(false));
-            f.Wait();
-            return f.Result;
         }
 
         public void Dispose()
